Apply @jolelievre comments
