+++ src/Hook/FeatureValue.php
@@ -32,6 +32,7 @@ use Tools;
 class FeatureValue extends AbstractHook
 {
     const AVAILABLE_HOOKS = [
+        'actionFeatureValueSave',
         'afterDeleteFeatureValue',
         'afterSaveFeatureValue',
         'displayFeatureValueForm',
@@ -130,5 +131,51 @@ class FeatureValue extends AbstractHook
         ]);
 
         return $this->module->render('feature_value_form.tpl');
+    }
+
+    /**
+     * After save feature
+     *
+     * @param array $params
+     */
+    public function actionFeatureValueSave(array $params)
+    {
+        if (empty($params['id_feature']) || Tools::getValue('layered_indexable') === false) {
+            return;
+        }
+
+        Db::getInstance()->execute(
+            'DELETE FROM ' . _DB_PREFIX_ . 'layered_indexable_feature
+            WHERE `id_feature` = ' . (int) $params['id_feature']
+        );
+        Db::getInstance()->execute(
+            'DELETE FROM ' . _DB_PREFIX_ . 'layered_indexable_feature_lang_value
+            WHERE `id_feature` = ' . (int) $params['id_feature']
+        );
+
+        Db::getInstance()->execute(
+            'INSERT INTO ' . _DB_PREFIX_ . 'layered_indexable_feature
+            (`id_feature`, `indexable`)
+            VALUES (' . (int) $params['id_feature'] . ', ' . (int) Tools::getValue('layered_indexable') . ')'
+        );
+
+        foreach (Language::getLanguages(false) as $language) {
+            $seoUrl = Tools::getValue('url_name_' . (int) $language['id_lang']);
+
+            if (empty($seoUrl)) {
+                $seoUrl = Tools::getValue('name_' . (int) $language['id_lang']);
+            }
+
+            Db::getInstance()->execute(
+                'INSERT INTO ' . _DB_PREFIX_ . 'layered_indexable_feature_lang_value
+                (`id_feature`, `id_lang`, `url_name`, `meta_title`)
+                VALUES (
+                ' . (int) $params['id_feature'] . ', ' . (int) $language['id_lang'] . ',
+                \'' . pSQL(Tools::link_rewrite($seoUrl)) . '\',
+                \'' . pSQL(Tools::getValue('meta_title_' . (int) $language['id_lang']), true) . '\')'
+            );
+        }
+
+        $this->module->invalidateLayeredFilterBlockCache();
     }
 }
